+++ src/etl_lib/__init__.py
@@ -1,4 +1,4 @@
 """
 Building blocks for ETL pipelines.
 """
-__version__ = "0.3.2"
+__version__ = "0.3.3"